requirements.txt
aiosqlite==0.20.0
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==24.3.0
azure-core==1.32.0
azure-identity==1.19.0
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
concurrent-log-handler==0.9.25
cryptography==44.0.0
dataclasses-json==0.6.7
Deprecated==1.2.15
dirtyjson==1.0.8
distro==1.9.0
docx==0.2.4
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iso8601==2.1.0
jiter==0.8.2
joblib==1.4.2
llama-index==0.10.43
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.43
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48.post4
llama-index-llms-azure-openai==0.1.9
llama-index-llms-openai==0.1.26
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
llamaindex-py-client==0.1.19
lxml==5.3.0
magic-filter==1.0.12
marshmallow==3.23.1
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
mypy-extensions==1.0.0
nbconvert==7.3.1
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
openai==1.57.4
pandas==2.2.3
pillow==11.0.0
portalocker==2.10.1
propcache==0.2.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
pypika-tortoise==0.5.0
pypdf==4.3.1
PyPDF2==3.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
striprtf==0.0.26
tenacity==8.5.0
tiktoken==0.8.0
tqdm==4.67.1
tortoise-orm==0.24.0
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.2.3
wrapt==1.17.0
yarl==1.18.3